test: 예약 결제 관련 기능 테스트 리팩토링
