+++ src/test/java/roomescape/reservation/application/BookingManageServiceTest.java
@@ -3,12 +3,8 @@ package roomescape.reservation.application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.event.ApplicationEvents;
-import org.springframework.test.context.event.RecordApplicationEvents;
 import roomescape.global.exception.ViolationException;
-import roomescape.payment.domain.ConfirmedPayment;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.event.ReservationSavedEvent;
 
 import java.util.List;
 
@@ -21,7 +17,6 @@ import static roomescape.TestFixture.MIA_RESERVATION_DATE;
 import static roomescape.reservation.domain.ReservationStatus.BOOKING;
 import static roomescape.reservation.domain.ReservationStatus.WAITING;
 
-@RecordApplicationEvents
 class BookingManageServiceTest extends ReservationServiceTest {
     @Autowired
     private BookingManageService bookingManageService;
@@ -31,9 +26,6 @@ class BookingManageServiceTest extends ReservationServiceTest {
 
     @Autowired
     private BookingQueryService bookingQueryService;
-
-    @Autowired
-    private ApplicationEvents events;
 
     @Test
     @DisplayName("예약을 생성한다.")
@@ -48,24 +40,6 @@ class BookingManageServiceTest extends ReservationServiceTest {
         assertSoftly(softly -> {
             softly.assertThat(createdReservation.getId()).isNotNull();
             softly.assertThat(createdReservation.getStatus()).isEqualTo(BOOKING);
-        });
-    }
-
-    @Test
-    @DisplayName("예약을 생성하고 결제 정보를 담은 이벤트를 발행한다.")
-    void createWithPayment() {
-        // given
-        Reservation reservation = MIA_RESERVATION(miaReservationTime, wootecoTheme, mia, BOOKING);
-        ConfirmedPayment confirmedPayment = new ConfirmedPayment("paymentKey", "orderId", 10);
-
-        // when
-        Reservation createdReservation = bookingManageService.createWithPayment(reservation, confirmedPayment);
-
-        // then
-        long count = events.stream(ReservationSavedEvent.class).count();
-        assertSoftly(softly -> {
-            softly.assertThat(createdReservation.getId()).isNotNull();
-            softly.assertThat(count).isEqualTo(1);
         });
     }
